#262 JWT `ttl` is overflowing

+++ api/src/main/java/com/stormpath/sdk/oauth/OauthGrantAuthenticationResult.java
@@ -65,6 +65,5 @@ public interface OauthGrantAuthenticationResult extends Oauth2AuthenticationResu
      * The lifetime in seconds of the access token. For example, the value "3600" denotes that the access token will expire one hour after the response was generated.
      * @return lifetime in seconds of the access token. For example, the value "3600" denotes that the access token will expire one hour after the response was generated.
      */
-    int getExpiresIn();
-
+    long getExpiresIn();
 }
